chore(pii): remove GLiNER/GPU image + add spaCy-skip fast path to CPU server
