--- apps/pii/requirements-gliner.txt
@@ -1,10 +1,0 @@
-# Extras for the opt-in GLiNER engine — installed ONLY in the `gliner`
-# Dockerfile target, on top of requirements.txt. Pinned for reproducible image
-# builds; bump deliberately. presidio-analyzer 2.2.362 requires
-# gliner >=0.2.13,<1.0.0.
-#
-# torch is pinned in the Dockerfile instead: the CPU and CUDA targets install
-# the same version from different wheel indexes.
-gliner==0.2.27
-transformers==5.3.0
-huggingface_hub==1.3.0
