+++ apps/pii/server.py
@@ -1,53 +1,251 @@
 """Combined Presidio REST service: analyzer + anonymizer on one port.
 
-Constructs one warm AnalyzerEngine (multi-language NLP + a native check-digit
-VIN recognizer) and one AnonymizerEngine at startup, exposing stock-compatible
-endpoints so a single PRESIDIO_URL serves both.
-
-NER engine selection (see engines.py):
-- PII_ENGINE=spacy (default): the 5 large spaCy models, unchanged behavior.
-- PII_ENGINE=gliner: one multilingual GLiNER model for PERSON/LOCATION/NRP/
-  DATE_TIME. The stock image ships both engines, so this is a pure env flip.
-  PII_DEVICE picks cpu/cuda (unset = auto-detect), PII_GLINER_MODEL overrides
-  the model id. The same code runs on CPU and GPU. Each uvicorn worker
-  (PII_WORKERS) loads its own GLiNER model copy — into GPU memory when on
-  cuda — so GPU deployments generally want PII_WORKERS=1 per GPU, unlike the
-  CPU/spacy path where workers scale with vCPUs.
+Constructs one warm AnalyzerEngine (5 large spaCy models for NER +
+regex/checksum pattern recognizers, incl. a native check-digit VIN recognizer)
+and one AnonymizerEngine at startup, exposing stock-compatible endpoints so a
+single PII_URL serves both.
 """
 
 import logging
-import os
 import time
 from typing import Any
 
-from engines import build_gliner_analyzer, build_spacy_analyzer
 from fastapi import FastAPI
-from presidio_analyzer import AnalyzerEngine, BatchAnalyzerEngine, RecognizerResult
+from presidio_analyzer import (
+    AnalyzerEngine,
+    BatchAnalyzerEngine,
+    Pattern,
+    PatternRecognizer,
+    RecognizerResult,
+)
+from presidio_analyzer.nlp_engine import NlpEngineProvider
+from presidio_analyzer.predefined_recognizers import (
+    AuAbnRecognizer,
+    AuAcnRecognizer,
+    AuMedicareRecognizer,
+    AuTfnRecognizer,
+    EsNieRecognizer,
+    EsNifRecognizer,
+    FiPersonalIdentityCodeRecognizer,
+    InAadhaarRecognizer,
+    InPanRecognizer,
+    InPassportRecognizer,
+    InVehicleRegistrationRecognizer,
+    InVoterRecognizer,
+    ItDriverLicenseRecognizer,
+    ItFiscalCodeRecognizer,
+    ItIdentityCardRecognizer,
+    ItPassportRecognizer,
+    ItVatCodeRecognizer,
+    PlPeselRecognizer,
+    SgFinRecognizer,
+    SgUenRecognizer,
+    UkNinoRecognizer,
+)
 from presidio_anonymizer import AnonymizerEngine
 from presidio_anonymizer.entities import OperatorConfig
 from pydantic import BaseModel
 
-PII_ENGINE = os.environ.get("PII_ENGINE", "spacy")
-if PII_ENGINE not in ("spacy", "gliner"):
-    raise ValueError(f"Invalid PII_ENGINE={PII_ENGINE!r}; expected 'spacy' or 'gliner'")
-# Empty/unset -> None -> auto-detect (cuda when torch sees a GPU, else cpu).
-PII_DEVICE = os.environ.get("PII_DEVICE") or None
-PII_GLINER_MODEL = os.environ.get("PII_GLINER_MODEL", "urchade/gliner_multi_pii-v1")
+# Languages served. Each needs its spaCy model installed in the image; the
+# es/it/pl/fi predefined recognizers (ES_NIF, IT_FISCAL_CODE, PL_PESEL, ...)
+# auto-load once their NLP engine is present.
+NLP_CONFIGURATION = {
+    "nlp_engine_name": "spacy",
+    "models": [
+        {"lang_code": "en", "model_name": "en_core_web_lg"},
+        {"lang_code": "es", "model_name": "es_core_news_lg"},
+        {"lang_code": "it", "model_name": "it_core_news_lg"},
+        {"lang_code": "pl", "model_name": "pl_core_news_lg"},
+        {"lang_code": "fi", "model_name": "fi_core_news_lg"},
+    ],
+}
+SUPPORTED_LANGUAGES = [m["lang_code"] for m in NLP_CONFIGURATION["models"]]
 
-# Propagates to uvicorn's root handler, so timing lands in the container log stream.
-logger = logging.getLogger("sim.pii")
+# spaCy pipeline components PII detection does not use. NER depends only on
+# `tok2vec` + `ner`; the parser (the most expensive stage), tagger, morphologizer,
+# attribute_ruler and lemmatizer are dead weight here. Disabling them is a ~2x
+# throughput win with NO change to the detected entities. The only side effect is
+# that Presidio's lemma-based context score-boosting weakens slightly — an
+# acceptable trade for the speed on this CPU-bound service.
+NER_ONLY_DISABLE = ("parser", "tagger", "morphologizer", "attribute_ruler", "lemmatizer")
+
+# Predefined recognizers Presidio ships but does NOT load into the default
+# registry — they must be added explicitly. Each carries its own
+# supported_language, so it fires under that language once its NLP model is
+# loaded. en: UK/AU/IN/SG locale ids; es/it/pl/fi: national ids.
+EXTRA_RECOGNIZERS = [
+    UkNinoRecognizer,
+    AuAbnRecognizer,
+    AuAcnRecognizer,
+    AuTfnRecognizer,
+    AuMedicareRecognizer,
+    InPanRecognizer,
+    InAadhaarRecognizer,
+    InVehicleRegistrationRecognizer,
+    InVoterRecognizer,
+    InPassportRecognizer,
+    SgFinRecognizer,
+    SgUenRecognizer,
+    EsNifRecognizer,
+    EsNieRecognizer,
+    ItFiscalCodeRecognizer,
+    ItDriverLicenseRecognizer,
+    ItVatCodeRecognizer,
+    ItPassportRecognizer,
+    ItIdentityCardRecognizer,
+    PlPeselRecognizer,
+    FiPersonalIdentityCodeRecognizer,
+]
+
+
+class VinRecognizer(PatternRecognizer):
+    """VIN (17 chars, A-Z/0-9 excluding I/O/Q) with ISO 3779 check-digit
+    validation (position 9). Validation makes accidental matches on arbitrary
+    17-char codes (request ids, SKUs, tokens) extremely unlikely. Some
+    non-North-American VINs omit the check digit and are skipped — an
+    intentional bias toward precision.
+    """
+
+    _TRANSLIT = {
+        **{str(d): d for d in range(10)},
+        "A": 1, "B": 2, "C": 3, "D": 4, "E": 5, "F": 6, "G": 7, "H": 8,
+        "J": 1, "K": 2, "L": 3, "M": 4, "N": 5, "P": 7, "R": 9,
+        "S": 2, "T": 3, "U": 4, "V": 5, "W": 6, "X": 7, "Y": 8, "Z": 9,
+    }
+    _WEIGHTS = [8, 7, 6, 5, 4, 3, 2, 10, 0, 9, 8, 7, 6, 5, 4, 3, 2]
+
+    def validate_result(self, pattern_text: str):
+        vin = pattern_text.upper()
+        if len(vin) != 17:
+            return False
+        try:
+            total = sum(self._TRANSLIT[c] * w for c, w in zip(vin, self._WEIGHTS))
+        except KeyError:
+            return False
+        check = total % 11
+        expected = "X" if check == 10 else str(check)
+        return vin[8] == expected
+
+
+def _register_common_recognizers(analyzer: AnalyzerEngine) -> None:
+    """Regex/checksum recognizers on top of spaCy NER + the Presidio defaults."""
+    # VIN is language-agnostic, so register it under every served language —
+    # a recognizer only fires for the language the caller routes to.
+    vin_pattern = Pattern(name="vin", regex=r"\b[A-HJ-NPR-Z0-9]{17}\b", score=0.7)
+    for language in SUPPORTED_LANGUAGES:
+        analyzer.registry.add_recognizer(
+            VinRecognizer(
+                supported_entity="VIN",
+                patterns=[vin_pattern],
+                context=["vin", "vehicle", "chassis"],
+                supported_language=language,
+            )
+        )
+    for recognizer_cls in EXTRA_RECOGNIZERS:
+        analyzer.registry.add_recognizer(recognizer_cls())
 
 
 def build_analyzer() -> AnalyzerEngine:
-    if PII_ENGINE == "gliner":
-        return build_gliner_analyzer(model_name=PII_GLINER_MODEL, device=PII_DEVICE)
-    return build_spacy_analyzer()
+    nlp_engine = NlpEngineProvider(nlp_configuration=NLP_CONFIGURATION).create_engine()
+    for nlp in getattr(nlp_engine, "nlp", {}).values():
+        for pipe in NER_ONLY_DISABLE:
+            if pipe in nlp.pipe_names:
+                nlp.disable_pipe(pipe)
+    analyzer = AnalyzerEngine(nlp_engine=nlp_engine, supported_languages=SUPPORTED_LANGUAGES)
+    _register_common_recognizers(analyzer)
+    return analyzer
 
 
-logger.info("building analyzer engine=%s device=%s", PII_ENGINE, PII_DEVICE or "auto")
+# Own handler at INFO. uvicorn configures only its own loggers, not the root, so a
+# bare getLogger propagates to a handler-less root at the default WARNING level and
+# every info() (the per-request timing lines) is silently dropped. Attach a stream
+# handler directly and stop propagation so timing lands in the container log stream
+# regardless of uvicorn's config or worker count.
+logger = logging.getLogger("sim.pii")
+logger.setLevel(logging.INFO)
+if not logger.handlers:
+    _log_handler = logging.StreamHandler()
+    _log_handler.setFormatter(logging.Formatter("%(levelname)s: %(name)s: %(message)s"))
+    logger.addHandler(_log_handler)
+    logger.propagate = False
+
+logger.info("building analyzer (spacy)")
 analyzer = build_analyzer()
 batch_analyzer = BatchAnalyzerEngine(analyzer_engine=analyzer)
 anonymizer = AnonymizerEngine()
+
+# The 4 spaCy-NER entity types. A request that names entities but none of these
+# needs no NLP pass — the regex/checksum recognizers match on the raw text. The
+# block-output redaction stage is restricted to exactly this case (regex-only).
+NER_ENTITIES = frozenset({"PERSON", "LOCATION", "NRP", "DATE_TIME"})
+
+# One blank NlpArtifacts per language, built once at startup. Passing these to
+# analyze() skips nlp_engine.process_text (the spaCy tok2vec+ner pass) entirely:
+# the pattern recognizers still match on the raw text, SpacyRecognizer is excluded
+# by the entity filter, and score_threshold is unset so detection is identical.
+# Only context-based score boosting (which needs real tokens) is unavailable — an
+# accepted trade for skipping NER on the hot block-output path. Read-only, so it
+# is safe to share across requests and workers.
+_BLANK_ARTIFACTS = {
+    language: analyzer.nlp_engine.process_text("", language)
+    for language in SUPPORTED_LANGUAGES
+}
+
+
+def _regex_only(entities: list[str] | None) -> bool:
+    """True when a request names entities and none are spaCy-NER, so the spaCy NLP
+    pass can be skipped."""
+    return bool(entities) and NER_ENTITIES.isdisjoint(entities)
+
+
+def _analyze_one(
+    text: str,
+    language: str,
+    entities: list[str] | None,
+    score_threshold: float | None,
+    return_decision_process: bool = False,
+):
+    # Regex-only requests reuse a blank NlpArtifacts to skip the spaCy NLP pass;
+    # otherwise analyze() computes artifacts (runs spaCy) as usual.
+    nlp_artifacts = _BLANK_ARTIFACTS.get(language) if _regex_only(entities) else None
+    return analyzer.analyze(
+        text=text,
+        language=language,
+        entities=entities or None,
+        score_threshold=score_threshold,
+        return_decision_process=return_decision_process,
+        nlp_artifacts=nlp_artifacts,
+    )
+
+
+def _analyze_many(
+    texts: list[str],
+    language: str,
+    entities: list[str] | None,
+    score_threshold: float | None,
+):
+    """Analyze many texts, skipping the spaCy pass for regex-only requests."""
+    if _regex_only(entities):
+        blank = _BLANK_ARTIFACTS.get(language)
+        return [
+            analyzer.analyze(
+                text=text,
+                language=language,
+                entities=entities,
+                score_threshold=score_threshold,
+                nlp_artifacts=blank,
+            )
+            for text in texts
+        ]
+    return list(
+        batch_analyzer.analyze_iterator(
+            texts=texts,
+            language=language,
+            entities=entities or None,
+            score_threshold=score_threshold,
+        )
+    )
+
 
 app = FastAPI(title="Sim Presidio", docs_url=None, redoc_url=None)
 
@@ -150,12 +348,12 @@ def supported_entities(language: str = "en") -> list[str]:
 @app.post("/analyze")
 def analyze(req: AnalyzeRequest) -> list[dict[str, Any]]:
     started = time.perf_counter()
-    results = analyzer.analyze(
-        text=req.text,
-        language=req.language,
-        entities=req.entities or None,
-        score_threshold=req.score_threshold,
-        return_decision_process=req.return_decision_process,
+    results = _analyze_one(
+        req.text,
+        req.language,
+        req.entities,
+        req.score_threshold,
+        req.return_decision_process,
     )
     logger.info(
         "analyze lang=%s chars=%d entities=%d duration_ms=%.1f",
@@ -171,12 +369,7 @@ def analyze(req: AnalyzeRequest) -> list[dict[str, Any]]:
 def analyze_batch(req: AnalyzeBatchRequest) -> list[list[dict[str, Any]]]:
     """Analyze many texts in one pass (spaCy nlp.pipe), returning one span list
     per input in request order — the batched counterpart to /analyze."""
-    results = batch_analyzer.analyze_iterator(
-        texts=req.texts,
-        language=req.language,
-        entities=req.entities or None,
-        score_threshold=req.score_threshold,
-    )
+    results = _analyze_many(req.texts, req.language, req.entities, req.score_threshold)
     return [[r.to_dict() for r in per_text] for per_text in results]
 
 
@@ -228,12 +421,7 @@ def redact(req: RedactRequest) -> dict[str, str]:
     anonymizer directly (no dict round-trip)."""
     started = time.perf_counter()
     operators = build_operators(req.anonymizers or req.operators)
-    results = analyzer.analyze(
-        text=req.text,
-        language=req.language,
-        entities=req.entities or None,
-        score_threshold=req.score_threshold,
-    )
+    results = _analyze_one(req.text, req.language, req.entities, req.score_threshold)
     text = (
         req.text
         if not results
@@ -261,14 +449,7 @@ def redact_batch(req: RedactBatchRequest) -> dict[str, list[str]]:
     texts that actually matched."""
     started = time.perf_counter()
     operators = build_operators(req.anonymizers or req.operators)
-    analyzed = list(
-        batch_analyzer.analyze_iterator(
-            texts=req.texts,
-            language=req.language,
-            entities=req.entities or None,
-            score_threshold=req.score_threshold,
-        )
-    )
+    analyzed = _analyze_many(req.texts, req.language, req.entities, req.score_threshold)
     masked: list[str] = []
     total_spans = 0
     for text, per_text in zip(req.texts, analyzed):
@@ -282,9 +463,11 @@ def redact_batch(req: RedactBatchRequest) -> dict[str, list[str]]:
             ).text
         )
     logger.info(
-        "redact_batch lang=%s texts=%d spans=%d duration_ms=%.1f",
+        "redact_batch lang=%s texts=%d entities=%s nlp=%s spans=%d duration_ms=%.1f",
         req.language,
         len(req.texts),
+        len(req.entities) if req.entities else "all",
+        "skip" if _regex_only(req.entities) else "full",
         total_spans,
         (time.perf_counter() - started) * 1000,
     )
