--- apps/pii/tests/conftest.py
@@ -1,5 +1,0 @@
-import sys
-from pathlib import Path
-
-# server.py / engines.py live one level up (repo: apps/pii, image: /app).
-sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
