+++ helm/sim/tests/pii_test.yaml
@@ -30,64 +30,18 @@ tests:
           path: spec.template.spec.containers[0].ports[0].containerPort
           value: 5001
 
-  - it: pii pod defaults to the spacy engine
-    template: deployment-pii.yaml
-    set:
-      pii.enabled: true
-    asserts:
-      - contains:
-          path: spec.template.spec.containers[0].env
-          content:
-            name: PII_ENGINE
-            value: "spacy"
-      - notContains:
-          path: spec.template.spec.containers[0].env
-          content:
-            name: PII_DEVICE
-            value: "cpu"
-
-  - it: pii.engine and pii.device render through to the container env
-    template: deployment-pii.yaml
-    set:
-      pii.enabled: true
-      pii.engine: gliner
-      pii.device: cuda
-    asserts:
-      - contains:
-          path: spec.template.spec.containers[0].env
-          content:
-            name: PII_ENGINE
-            value: "gliner"
-      - contains:
-          path: spec.template.spec.containers[0].env
-          content:
-            name: PII_DEVICE
-            value: "cuda"
-
-  - it: user-set pii.env cannot override the chart-owned engine keys
+  - it: pii.env renders through to the container env
     template: deployment-pii.yaml
     set:
       pii.enabled: true
       pii.env:
-        PII_ENGINE: evil
-        PII_DEVICE: evil
-        PII_GLINER_MODEL: acme/custom-model
+        PII_WORKERS: "4"
     asserts:
-      - notContains:
-          path: spec.template.spec.containers[0].env
-          content:
-            name: PII_ENGINE
-            value: "evil"
-      - notContains:
-          path: spec.template.spec.containers[0].env
-          content:
-            name: PII_DEVICE
-            value: "evil"
       - contains:
           path: spec.template.spec.containers[0].env
           content:
-            name: PII_GLINER_MODEL
-            value: "acme/custom-model"
+            name: PII_WORKERS
+            value: "4"
 
   - it: app pod gets chart-computed PII_URL pointing at the in-cluster service
     template: deployment-app.yaml
