+++ helm/sim/values.yaml
@@ -836,18 +836,6 @@ pii:
     digest: ""       # sha256: pin overrides tag
     pullPolicy: IfNotPresent
 
-  # NER engine: "spacy" (default) or "gliner" (opt-in zero-shot transformer NER
-  # for PERSON/LOCATION/NRP/DATE_TIME; regex/checksum recognizers are identical
-  # on both engines). The image ships both engines, so this is a pure env flip
-  # — no image change needed. Note: gliner on CPU is orders of magnitude slower
-  # than spacy; it is intended for GPU nodes.
-  engine: "spacy"
-
-  # Torch device for the gliner engine: "cpu", "cuda", or "cuda:N". Empty
-  # auto-detects (cuda when a GPU is visible, else cpu). GPU resource requests
-  # (nvidia.com/gpu) are not wired yet — GPU scheduling is an infra follow-up.
-  device: ""
-
   # Number of replicas
   replicaCount: 1
